BrowserSim: refactoring skins
