+++ plugins/org.jboss.tools.vpe.browsersim/src/org/jboss/tools/vpe/browsersim/ui/skin/NativeSkin.java
@@ -234,20 +234,14 @@ public class NativeSkin implements BrowserSimSkin {
 	public void setOrientationAndSize(int orientation, Point browserSize, ResizableSkinSizeAdvisor sizeAdvisor) {
 		GridData data = (GridData) browser.getLayoutData();
 		
-		int shellWidthHint = SWT.DEFAULT;
 		if (browserSize.x != Device.DEFAULT_SIZE) {
 			data.widthHint = browserSize.x;
 		} 
-//		else if (data.widthHint == SWT.DEFAULT) {
-//			shellWidthHint = maximumShellSize.x;
-//		}
-		int shellHeightHint = SWT.DEFAULT;
+
 		if (browserSize.y != Device.DEFAULT_SIZE) {
 			data.heightHint =  browserSize.y;
 		}
-//		else if (data.heightHint == SWT.DEFAULT) {
-//			shellHeightHint = maximumShellSize.y;
-//		}
+
 		Point prefferedShellSize = shell.computeSize(SWT.DEFAULT, SWT.DEFAULT);
 		
 		Point shellSize = sizeAdvisor.checkWindowSize(orientation, browserSize, prefferedShellSize);
