+++ plugins/org.jboss.tools.vpe.browsersim/src/org/jboss/tools/vpe/browsersim/ui/skin/ios/AppleIPadMiniResizableSkin.java
@@ -17,39 +17,40 @@ import org.eclipse.swt.layout.FormData;
 import org.eclipse.swt.widgets.Composite;
 import org.jboss.tools.vpe.browsersim.ui.skin.DeviceComposite;
 import org.jboss.tools.vpe.browsersim.ui.skin.ImageDescriptor;
+import org.jboss.tools.vpe.browsersim.ui.skin.ResizableSkin;
 
 /**
  * @author Ilya Buziuk (ibuziuk)
  */
 @SuppressWarnings("nls")
-public class AppleIPadMiniResizableSkin extends AppleIPhone3ResizableSkin {
+public class AppleIPadMiniResizableSkin extends ResizableSkin {
 	public static final String IPAD_MINI_SKIN_ID = "iPad mini";
+	private static final int[] VISIBLE_REGION_VERTICAL = { 37, 0, 341, 0, 342, 1, 346, 1, 347, 2, 349, 2, 350, 3, 352,
+		3, 353, 4, 354, 4, 355, 5, 356, 5, 358, 7, 359, 7, 361, 9, 362, 9, 368, 15, 368, 16, 371, 19, 371, 20, 372,
+		21, 372, 22, 373, 23, 373, 24, 374, 25, 374, 26, 375, 27, 375, 29, 376, 30, 376, 33, 377, 34, 377, 576,
+		376, 577, 376, 579, 375, 580, 375, 582, 374, 583, 374, 585, 373, 586, 373, 587, 371, 589, 371, 590, 369,
+		592, 369, 593, 361, 601, 360, 601, 358, 603, 357, 603, 355, 605, 353, 605, 352, 606, 351, 606, 350, 607,
+		348, 607, 347, 608, 344, 608, 343, 609, 34, 609, 33, 608, 30, 608, 29, 607, 27, 607, 26, 606, 25, 606, 24,
+		605, 23, 605, 22, 604, 21, 604, 20, 603, 19, 603, 16, 600, 15, 600, 9, 594, 9, 593, 7, 591, 7, 590, 5, 588,
+		5, 587, 4, 586, 4, 585, 3, 584, 3, 582, 2, 581, 2, 579, 1, 578, 1, 574, 0, 573, 0, 36, 1, 35, 1, 31, 2, 30,
+		2, 28, 3, 27, 3, 26, 4, 25, 4, 24, 5, 23, 5, 22, 6, 21, 6, 20, 7, 19, 7, 18, 12, 13, 12, 12, 13, 12, 17, 8,
+		18, 8, 20, 6, 21, 6, 22, 5, 23, 5, 24, 4, 25, 4, 26, 3, 28, 3, 29, 2, 31, 2, 32, 1, 36, 1 };
+	private static final int[] VISIBLE_REGION_HORIZONTAL = { 34, 0, 576, 0, 577, 1, 579, 1, 580, 2, 582, 2, 583, 3,
+		585, 3, 586, 4, 587, 4, 589, 6, 590, 6, 592, 8, 593, 8, 601, 16, 601, 17, 603, 19, 603, 20, 605, 22, 605,
+		24, 606, 25, 606, 26, 607, 27, 607, 29, 608, 30, 608, 33, 609, 34, 609, 343, 608, 344, 608, 347, 607, 348,
+		607, 350, 606, 351, 606, 352, 605, 353, 605, 354, 604, 355, 604, 356, 603, 357, 603, 358, 600, 361, 600,
+		362, 594, 368, 593, 368, 591, 370, 590, 370, 588, 372, 587, 372, 586, 373, 585, 373, 584, 374, 582, 374,
+		581, 375, 579, 375, 578, 376, 574, 376, 573, 377, 36, 377, 35, 376, 31, 376, 30, 375, 28, 375, 27, 374, 26,
+		374, 25, 373, 24, 373, 23, 372, 22, 372, 21, 371, 20, 371, 19, 370, 18, 370, 13, 365, 12, 365, 12, 364, 8,
+		360, 8, 359, 6, 357, 6, 356, 5, 355, 5, 354, 4, 353, 4, 352, 3, 351, 3, 349, 2, 348, 2, 346, 1, 345, 1,
+		341, 0, 340, 0, 36, 1, 35, 1, 31, 2, 30, 2, 28, 3, 27, 3, 25, 4, 24, 4, 23, 5, 22, 5, 21, 7, 19, 7, 18, 9,
+		16, 9, 15, 15, 9, 16, 9, 19, 6, 20, 6, 21, 5, 22, 5, 23, 4, 24, 4, 25, 3, 26, 3, 27, 2, 29, 2, 30, 1, 33, 1 };
 	private static final Point NORMAL_SCREEN_SIZE = new Point(318, 450);
 	private static final Point NORMAL_SKIN_SIZE = new Point(378, 610);
-	private static final Point VERTICAL_BORDERS_SIZE = new Point(NORMAL_SKIN_SIZE.x - NORMAL_SCREEN_SIZE.x, NORMAL_SKIN_SIZE.y - NORMAL_SCREEN_SIZE.y);
-	private static final Point HORIZONTAL_BORDERS_SIZE = new Point(VERTICAL_BORDERS_SIZE.y, VERTICAL_BORDERS_SIZE.x);
-	private static final int[] VISIBLE_REGION_VERTICAL = { 37, 0, 341, 0, 342, 1, 346, 1, 347, 2, 349, 2, 350, 3, 352,
-			3, 353, 4, 354, 4, 355, 5, 356, 5, 358, 7, 359, 7, 361, 9, 362, 9, 368, 15, 368, 16, 371, 19, 371, 20, 372,
-			21, 372, 22, 373, 23, 373, 24, 374, 25, 374, 26, 375, 27, 375, 29, 376, 30, 376, 33, 377, 34, 377, 576,
-			376, 577, 376, 579, 375, 580, 375, 582, 374, 583, 374, 585, 373, 586, 373, 587, 371, 589, 371, 590, 369,
-			592, 369, 593, 361, 601, 360, 601, 358, 603, 357, 603, 355, 605, 353, 605, 352, 606, 351, 606, 350, 607,
-			348, 607, 347, 608, 344, 608, 343, 609, 34, 609, 33, 608, 30, 608, 29, 607, 27, 607, 26, 606, 25, 606, 24,
-			605, 23, 605, 22, 604, 21, 604, 20, 603, 19, 603, 16, 600, 15, 600, 9, 594, 9, 593, 7, 591, 7, 590, 5, 588,
-			5, 587, 4, 586, 4, 585, 3, 584, 3, 582, 2, 581, 2, 579, 1, 578, 1, 574, 0, 573, 0, 36, 1, 35, 1, 31, 2, 30,
-			2, 28, 3, 27, 3, 26, 4, 25, 4, 24, 5, 23, 5, 22, 6, 21, 6, 20, 7, 19, 7, 18, 12, 13, 12, 12, 13, 12, 17, 8,
-			18, 8, 20, 6, 21, 6, 22, 5, 23, 5, 24, 4, 25, 4, 26, 3, 28, 3, 29, 2, 31, 2, 32, 1, 36, 1 };
-	private static final int[] VISIBLE_REGION_HORIZONTAL = { 34, 0, 576, 0, 577, 1, 579, 1, 580, 2, 582, 2, 583, 3,
-			585, 3, 586, 4, 587, 4, 589, 6, 590, 6, 592, 8, 593, 8, 601, 16, 601, 17, 603, 19, 603, 20, 605, 22, 605,
-			24, 606, 25, 606, 26, 607, 27, 607, 29, 608, 30, 608, 33, 609, 34, 609, 343, 608, 344, 608, 347, 607, 348,
-			607, 350, 606, 351, 606, 352, 605, 353, 605, 354, 604, 355, 604, 356, 603, 357, 603, 358, 600, 361, 600,
-			362, 594, 368, 593, 368, 591, 370, 590, 370, 588, 372, 587, 372, 586, 373, 585, 373, 584, 374, 582, 374,
-			581, 375, 579, 375, 578, 376, 574, 376, 573, 377, 36, 377, 35, 376, 31, 376, 30, 375, 28, 375, 27, 374, 26,
-			374, 25, 373, 24, 373, 23, 372, 22, 372, 21, 371, 20, 371, 19, 370, 18, 370, 13, 365, 12, 365, 12, 364, 8,
-			360, 8, 359, 6, 357, 6, 356, 5, 355, 5, 354, 4, 353, 4, 352, 3, 351, 3, 349, 2, 348, 2, 346, 1, 345, 1,
-			341, 0, 340, 0, 36, 1, 35, 1, 31, 2, 30, 2, 28, 3, 27, 3, 25, 4, 24, 4, 23, 5, 22, 5, 21, 7, 19, 7, 18, 9,
-			16, 9, 15, 15, 9, 16, 9, 19, 6, 20, 6, 21, 5, 22, 5, 23, 4, 24, 4, 25, 3, 26, 3, 27, 2, 29, 2, 30, 1, 33, 1 };
+	
+	private static final int CORNERS_SIZE = 40;		
+	
 	private static final IPhoneSkinDescriptor VERTICAL_DESCRIPTOR;
-	private static final int CORNERS_SIZE = 40;
 	static {
 		String bd = "ios/ipad/mini/vertical/";
 		ImageDescriptor iOsDescriptor = new ImageDescriptor(null, 5, 3, SWT.VERTICAL | SWT.HORIZONTAL);
@@ -113,10 +114,8 @@ public class AppleIPadMiniResizableSkin extends AppleIPhone3ResizableSkin {
 		HORIZONTAL_DESCRIPTOR = new IPhoneSkinDescriptor(bodyGridSize, bodyGridImageDescriptors, iOsDescriptor, CORNERS_SIZE, backButtonDescriptor, forwardButtonDescriptor);
 	}
 	
-	@Override
-	protected Point getBordersSize(boolean vertical) {
-		Point bordersSize = vertical ? VERTICAL_BORDERS_SIZE : HORIZONTAL_BORDERS_SIZE;
-		return bordersSize;
+	public AppleIPadMiniResizableSkin() {
+		super(VISIBLE_REGION_HORIZONTAL, VISIBLE_REGION_VERTICAL, NORMAL_SCREEN_SIZE, NORMAL_SKIN_SIZE);
 	}
 
 	@Override
@@ -128,15 +127,6 @@ public class AppleIPadMiniResizableSkin extends AppleIPhone3ResizableSkin {
 			skinDescriptor = HORIZONTAL_DESCRIPTOR;
 		}
 		return new AppleIPadComposite(parent, skinDescriptor);
-	}
-	
-	@Override
-	protected int[] getNormalRegion(boolean vertical) {
-		if (vertical) {
-			return VISIBLE_REGION_VERTICAL;
-		} else {
-			return VISIBLE_REGION_HORIZONTAL;
-		}
 	}
 
 	@Override
